fix: improve arbigent performance and data type handling
- Add resource limits to mosaiClassifier_calc_probs_arbigent rule
- Change arbigent HDF storage from int8 to int32 for larger count values
- Update watson_crick.py arbigent utilities

This prevents data overflow issues with high read counts and ensures
proper resource allocation for arbigent analysis.

+++ workflow/scripts/arbigent_utils/create_hdf.py
@@ -84,8 +84,8 @@ def convert_mapping_counts(raw_counts, process_chrom):
 
             if chrom != last_chrom:
                 with pd.HDFStore(hdf_file, "a") as hdf:
-                    hdf.put(os.path.join(last_chrom, "correct"), pd.Series(correct_counts, dtype=np.int8), format="fixed")
-                    hdf.put(os.path.join(last_chrom, "incorrect"), pd.Series(incorrect_counts, dtype=np.int8), format="fixed")
+                    hdf.put(os.path.join(last_chrom, "correct"), pd.Series(correct_counts, dtype=np.int32), format="fixed")
+                    hdf.put(os.path.join(last_chrom, "incorrect"), pd.Series(incorrect_counts, dtype=np.int32), format="fixed")
 
                 correct_counts = []
                 incorrect_counts = []
@@ -109,8 +109,8 @@ def convert_mapping_counts(raw_counts, process_chrom):
     # dump last
     if correct_counts:
         with pd.HDFStore(hdf_file, "a") as hdf:
-            hdf.put(os.path.join(last_chrom, "correct"), pd.Series(correct_counts, dtype=np.int8), format="fixed")
-            hdf.put(os.path.join(last_chrom, "incorrect"), pd.Series(incorrect_counts, dtype=np.int8), format="fixed")
+            hdf.put(os.path.join(last_chrom, "correct"), pd.Series(correct_counts, dtype=np.int32), format="fixed")
+            hdf.put(os.path.join(last_chrom, "incorrect"), pd.Series(incorrect_counts, dtype=np.int32), format="fixed")
 
     # return hdf_file
